cleaning the gradient set up
